+++ src/app/api/strategy/test-positions/route.ts
@@ -11,5 +11,5 @@ export async function GET(req: NextRequest) {
     } catch (error) {
         console.log('error:', error);
         return NextResponse.json({ error: 'Failed to fetch data' }, { status: 500 });
-    } 
+    }
 }