+++ src/components/alpaca/dashboard/dashboard.tsx
@@ -1,22 +1,17 @@
 "use client";
 import { useState } from "react";
-import TestPositions from "../strategy-test/test-results";
-import StrategySettingsForm from "../strategy-test/strategy-settings-form";
-import StrategyList from "../strategy-test/strategy-list";
-import { BacktestSettings } from "@/models/strategy/test-settings";
+import { StrategySettingsModel } from "@/models/strategy/strategy-settings-model";
 import AlpacaAccount from "./alpaca-account";
 import Strategies from "./strategies";
 import AlpacaPositions from "./alpaca-positions";
 import StrategyExecution from "./strategy-execution";
 import StrategyTestResult from "./strategy-test-result";
 
-
-
 const Dashboard: React.FC = () => {
 
-    const [testSettings, setTestSettings] = useState<BacktestSettings>({} as BacktestSettings);
+    const [testSettings, setTestSettings] = useState<StrategySettingsModel>({} as StrategySettingsModel);
 
-    const updateComponents = (e: BacktestSettings) => {
+    const updateComponents = (e: StrategySettingsModel) => {
         setTestSettings(e);
     }
 
@@ -30,14 +25,12 @@ const Dashboard: React.FC = () => {
                     <Strategies showResult={updateComponents} />
                 </div>
                 <div className="flex-1">
-                    <StrategyTestResult test={testSettings} />
+                    <StrategyTestResult strat={testSettings} />
                 </div>
                 <div className="flex-1">
                     <div className="h-[30%] bg-white">
-
                         <StrategyExecution />
                         <AlpacaPositions />
-
                     </div>
                 </div>
             </div>