+++ src/components/alpaca/strategy-test/strategy-settings-form.tsx
@@ -7,10 +7,11 @@ import SubmitButton from "@/components/common/buttons/submit-button";
 import { format } from 'date-fns';
 import { BreakoutPeriod, StopLossStrategy } from "@/models/strategy/enums";
 import { useDictionary } from "@/provider/dictionary-provider";
-import { useForm } from "react-hook-form";
-
 import WidgetButton from "@/components/common/buttons/widget-button";
 import CheckboxSlate from "@/components/common/checkbox/checkbox-slate";
+import { basicFetch } from "@/app/lib/fetchFunctions";
+import { AlpacaAssetModel } from "@/models/alpaca/alpaca-asset-model";
+import CircularLoader from "@/components/common/loader";
 
 enum StrategySettingsFormState {
     None,
@@ -24,14 +25,21 @@ enum StrategySettingsFormState {
 const StrategySettingsForm = () => {
     const dictionary = useDictionary();
     const [formState, setFormState] = useState<StrategySettingsFormState>(StrategySettingsFormState.None);
+    const [assets, setAssets] = useState<AlpacaAssetModel[]>([]);
 
     const [state, storeAction] = useFormState<any, FormData>(createAlpacaBacktest, { message: '', success: false, errors: {} });
-    const [pending, setPending] = useState(false);
+    const [pending, setPending] = useState<boolean>(false);
+    const [loading, setLoading] = useState<boolean>(true);
 
     const [strategy, setStrategy] = useState<string>('0');
     const [stopLossStrategy, setStopLossStrategy] = useState<string>('0');
     const today = new Date()
     const formattedDate = format(today, 'yyyy-MM-dd');
+
+
+    useEffect(() => {
+        getAssets();
+    }, []);
 
 
     useEffect(() => {
@@ -47,15 +55,17 @@ const StrategySettingsForm = () => {
 
     }, [state]);
 
-    const handleSubmit = (e: any) => {
+    const getAssets = async () => {
+        const assets = await basicFetch<any>(`/api/alpaca/assets`);
+        setAssets(assets);
+    }
+
+    const handleSubmit = (e: boolean) => {
         setPending(e);
         console.log('pending:', pending);
     }
 
-    const handleFocus = (fieldName: any) => {
-
-
-    };
+    const handleFocus = (fieldName: any) => {   };
 
     return (
         <div className="component-container  overflow-hidden">
@@ -69,7 +79,7 @@ const StrategySettingsForm = () => {
                         <table className="w-full border-collapse">
                             <tbody>
                                 <tr>
-                                    <td className="pb-1"><label>Backtest Name</label></td>
+                                    <td className="pb-1  w-[30%]"><label>Backtest Name</label></td>
                                     <td className="pb-1">
                                         <input type="text" name="name" className="border border-slate-400 w-full p-1" defaultValue="test" disabled={pending} />
                                         <div className="error-message">{firstOrDefault(state?.errors?.name, '')}</div>
@@ -91,20 +101,23 @@ const StrategySettingsForm = () => {
                                             <td className="pb-1"><label>Symbol</label></td>
                                             <td className="pb-1">
                                                 <select name="symbol" className="border border-slate-400 w-full p-1" title="Symbol" disabled={pending}>
-                                                    <option value="SPY">SPY</option>
-                                                    <option value="TSLA">TSLA</option>
+                                                    {assets.map((asset) => (
+                                                        <option key={asset.assetId} value={asset.symbol}>
+                                                            <div className="truncate w-10">{asset.symbol} {asset.name}</div>
+                                                        </option>
+                                                    ))}
                                                 </select>
                                             </td>
                                         </tr>
                                         <tr>
-                                            <td className="pb-1"><label>Start Date</label></td>
+                                            <td className="pb-1"><label>Start Backtest</label></td>
                                             <td className="pb-1">
                                                 <input type="date" name="startDate" className="border border-slate-400 w-full p-1" defaultValue="2024-01-01" disabled={pending} />
                                                 <div className="error-message">{firstOrDefault(state?.errors?.startDate, '')}</div>
                                             </td>
                                         </tr>
                                         <tr>
-                                            <td className="pb-1"><label>End Date</label></td>
+                                            <td className="pb-1"><label>End Backtest</label></td>
                                             <td className="pb-1">
                                                 <input type="date" name="endDate" className="border border-slate-400 w-full p-1" defaultValue={formattedDate} disabled={pending} />
                                                 <div className="error-message">{firstOrDefault(state?.errors?.endDate, '')}</div>