+++ src/models/alpaca/alpaca-asset-model.ts
@@ -1,0 +1,6 @@
+export interface AlpacaAssetModel {
+    assetId: string;
+    symbol: string;
+    name: string;
+    marketCloseTime: boolean;
+}